Added Parser class, changed AnaPyzerAnalyzer class so it's formatted more like our other classes.  Refactored controller method calls to account for new structure
